user service interface

+++ src/main/java/com/expensetracker/services/UserService.java
@@ -14,5 +14,5 @@ public interface UserService {
     void deleteById(long id);
     void deleteByUsername(String username);
     void deleteAll();
-    long setIncome(String username, long income);
+    Double setIncome(String username, double income);
 }
